Remove references to FilterOptions
There's no reasonable way to retrieve the set of all values for all parameters for each subject area.
For example, the legal values for `meet_locations` depends on the result of the initial query for all root sections. It doesn't make sense to utilize time to retrieve all sections and their children and parse all the data, only to return a smaller subset.

Filtering should instead be done by the implementing library, if desired.